refactor: 주문과 테이블의 의존성을 역전시켜라
