+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -15,10 +15,10 @@ import kitchenpos.core.exception.BadRequestException;
 import kitchenpos.core.exception.CannotUpdateException;
 import kitchenpos.core.exception.ExceptionType;
 import kitchenpos.core.exception.NotFoundException;
-import kitchenpos.order.application.OrderTableService;
+import kitchenpos.table.application.OrderTableService;
 import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.table.application.OrderTableServiceImpl;
+import kitchenpos.order.application.OrderTableServiceImpl;
 import kitchenpos.table.application.TableService;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
